+++ deploy-scripts/build-devnet.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-anchor build -- --no-default-features
+anchor build -- --no-default-features --features no-entrypoint