+++ programs/drift/src/math/funding/tests.rs
@@ -1,7 +1,7 @@
 use crate::controller::funding::update_funding_rate;
 use crate::controller::repeg::_update_amm;
 use crate::math::helpers::on_the_hour_update;
-use crate::math::oracle::block_operation;
+use crate::math::oracle::{block_operation, OracleValidity};
 
 use crate::math::constants::{
     AMM_RESERVE_PRECISION, ONE_HOUR_I128, PRICE_PRECISION, PRICE_PRECISION_U64, QUOTE_PRECISION,
@@ -12,7 +12,7 @@ use std::cmp::min;
 use crate::test_utils::get_pyth_price;
 
 // use crate::create_anchor_account_info;
-use crate::state::oracle::HistoricalOracleData;
+use crate::state::oracle::{HistoricalOracleData, MMOraclePriceData};
 use crate::state::oracle_map::OracleMap;
 use crate::state::perp_market::{ContractTier, PerpMarket, AMM};
 use crate::state::state::{OracleGuardRails, State, ValidityGuardRails};
@@ -653,6 +653,13 @@ fn unsettled_funding_pnl() {
     )
     .unwrap();
     let oracle_price_data = oracle_map.get_price_data(&market.oracle_id()).unwrap();
+    let mm_oracle_price_data = MMOraclePriceData::new(
+        oracle_price_data.price,
+        oracle_price_data.delay + 1,
+        OracleValidity::default(),
+        *oracle_price_data,
+    )
+    .unwrap();
 
     assert_eq!(time_until_next_update, 0);
     let block_funding_rate_update = block_operation(
@@ -669,7 +676,7 @@ fn unsettled_funding_pnl() {
     now += 3600;
     slot += 3600 * 2;
 
-    let cost = _update_amm(&mut market, oracle_price_data, &state, now, slot).unwrap();
+    let cost = _update_amm(&mut market, &mm_oracle_price_data, &state, now, slot).unwrap();
     assert_eq!(cost, 0);
     assert_eq!(market.amm.last_update_slot, slot);
     assert_eq!(market.amm.last_mark_price_twap, 50000000);