+++ sdk/src/dlob/NodeList.ts
@@ -1,5 +1,11 @@
-import { isVariant, MarketTypeStr, Order, ProtectedMakerParams } from '..';
+import {
+	isVariant,
+	MarketTypeStr,
+	Order,
+	ProtectedMakerParams,
+} from '../types';
 import { createNode, DLOBNode, DLOBNodeMap } from './DLOBNode';
+import { BN } from '@coral-xyz/anchor';
 
 export type SortDirection = 'asc' | 'desc';
 
@@ -37,7 +43,8 @@ export class NodeList<NodeType extends keyof DLOBNodeMap>
 		marketType: MarketTypeStr,
 		userAccount: string,
 		isProtectedMaker: boolean,
-		protectedMakerParamsMap?: ProtectedMakerParams
+		protectedMakerParamsMap?: ProtectedMakerParams,
+		baseAssetAmount?: BN
 	): void {
 		if (!isVariant(order.status, 'open')) {
 			return;
@@ -48,7 +55,8 @@ export class NodeList<NodeType extends keyof DLOBNodeMap>
 			order,
 			userAccount,
 			isProtectedMaker,
-			protectedMakerParamsMap
+			protectedMakerParamsMap,
+			baseAssetAmount
 		);
 
 		const orderSignature = getOrderSignature(order.orderId, userAccount);