+++ sdk/src/dlob/orderBookLevels.ts
@@ -1,27 +1,30 @@
+import { BN } from '@coral-xyz/anchor';
 import {
 	BASE_PRECISION,
-	BN,
+	QUOTE_PRECISION,
+	ZERO,
+	PRICE_PRECISION,
+	AMM_TO_QUOTE_PRECISION_RATIO,
+} from '../constants/numericConstants';
+import {
 	calculateAmmReservesAfterSwap,
 	calculateMarketOpenBidAsk,
 	calculateQuoteAssetAmountSwapped,
 	calculateSpreadReserves,
 	calculateUpdatedAMM,
-	DLOBNode,
-	isOperationPaused,
+} from '../math/amm';
+import { DLOBNode } from './DLOBNode';
+import { isOperationPaused } from '../math/exchangeStatus';
+import {
 	isVariant,
-	OraclePriceData,
 	PerpMarketAccount,
 	PerpOperation,
 	PositionDirection,
-	QUOTE_PRECISION,
-	standardizePrice,
 	SwapDirection,
-	ZERO,
-	PRICE_PRECISION,
-	AMM_TO_QUOTE_PRECISION_RATIO,
-	standardizeBaseAssetAmount,
-} from '..';
+} from '../types';
+import { OraclePriceData } from '../oracles/types';
 import { PublicKey } from '@solana/web3.js';
+import { standardizeBaseAssetAmount, standardizePrice } from '../math/orders';
 
 type liquiditySource =
 	| 'serum'
@@ -75,22 +78,28 @@ export const MAJORS_TOP_OF_BOOK_QUOTE_AMOUNTS = [
 	new BN(50000).mul(QUOTE_PRECISION),
 ];
 
+const INDICATIVE_QUOTES_PUBKEY = 'inDNdu3ML4vG5LNExqcwuCQtLcCU8KfK5YM2qYV3JJz';
+
 /**
  * Get an {@link Generator<L2Level>} generator from a {@link Generator<DLOBNode>}
  * @param dlobNodes e.g. {@link DLOB#getRestingLimitAsks} or {@link DLOB#getRestingLimitBids}
  * @param oraclePriceData
  * @param slot
  */
-const INDICATIVE_QUOTES_PUBKEY = 'inDNdu3ML4vG5LNExqcwuCQtLcCU8KfK5YM2qYV3JJz';
 export function* getL2GeneratorFromDLOBNodes(
 	dlobNodes: Generator<DLOBNode>,
 	oraclePriceData: OraclePriceData,
 	slot: number
 ): Generator<L2Level> {
 	for (const dlobNode of dlobNodes) {
-		const size = dlobNode.order.baseAssetAmount.sub(
+		const size = dlobNode.baseAssetAmount.sub(
 			dlobNode.order.baseAssetAmountFilled
 		) as BN;
+
+		if (size.lte(ZERO)) {
+			continue;
+		}
+
 		yield {
 			size,
 			price: dlobNode.getPrice(oraclePriceData, slot),