+++ sdk/src/driftClient.ts
@@ -78,6 +78,7 @@ import {
 	PublicKey,
 	Signer,
 	SystemProgram,
+	SYSVAR_CLOCK_PUBKEY,
 	SYSVAR_INSTRUCTIONS_PUBKEY,
 	Transaction,
 	TransactionInstruction,
@@ -122,6 +123,8 @@ import { TxSender, TxSigAndSlot } from './tx/types';
 import {
 	BASE_PRECISION,
 	GOV_SPOT_MARKET_INDEX,
+	ONE,
+	PERCENTAGE_PRECISION,
 	PRICE_PRECISION,
 	QUOTE_PRECISION,
 	QUOTE_SPOT_MARKET_INDEX,
@@ -177,7 +180,10 @@ import { WormholeCoreBridgeSolana } from '@pythnetwork/pyth-solana-receiver/lib/
 import { PythSolanaReceiver } from '@pythnetwork/pyth-solana-receiver/lib/idl/pyth_solana_receiver';
 import { getFeedIdUint8Array, trimFeedId } from './util/pythOracleUtils';
 import { createMinimalEd25519VerifyIx } from './util/ed25519Utils';
-import { isVersionedTransaction } from './tx/utils';
+import {
+	createNativeInstructionDiscriminatorBuffer,
+	isVersionedTransaction,
+} from './tx/utils';
 import pythSolanaReceiverIdl from './idl/pyth_solana_receiver.json';
 import { asV0Tx, PullFeed, AnchorUtils } from '@switchboard-xyz/on-demand';
 import { gprcDriftClientAccountSubscriber } from './accounts/grpcDriftClientAccountSubscriber';
@@ -186,6 +192,7 @@ import { Slothash } from './slot/SlothashSubscriber';
 import { getOracleId } from './oracles/oracleId';
 import { SignedMsgOrderParams } from './types';
 import { sha256 } from '@noble/hashes/sha256';
+import { getOracleConfidenceFromMMOracleData } from './oracles/utils';
 
 type RemainingAccountParams = {
 	userAccounts: UserAccount[];
@@ -439,7 +446,8 @@ export class DriftClient {
 					resubTimeoutMs: config.accountSubscription?.resubTimeoutMs,
 					logResubMessages: config.accountSubscription?.logResubMessages,
 				},
-				config.accountSubscription?.commitment
+				config.accountSubscription?.commitment,
+				config.accountSubscription?.perpMarketAccountSubscriber
 			);
 		}
 		this.eventEmitter = this.accountSubscriber.eventEmitter;
@@ -8480,9 +8488,52 @@ export class DriftClient {
 	}
 
 	public getOracleDataForPerpMarket(marketIndex: number): OraclePriceData {
-		return this.accountSubscriber.getOraclePriceDataAndSlotForPerpMarket(
-			marketIndex
-		).data;
+		const perpMarket = this.getPerpMarketAccount(marketIndex);
+		const isMMOracleActive = !perpMarket.amm.mmOracleSlot.eq(ZERO);
+		return {
+			...this.accountSubscriber.getOraclePriceDataAndSlotForPerpMarket(
+				marketIndex
+			).data,
+			isMMOracleActive,
+		};
+	}
+
+	public getMMOracleDataForPerpMarket(marketIndex: number): OraclePriceData {
+		const perpMarket = this.getPerpMarketAccount(marketIndex);
+		const oracleData = this.getOracleDataForPerpMarket(marketIndex);
+		const stateAccountAndSlot = this.accountSubscriber.getStateAccountAndSlot();
+		const pctDiff = perpMarket.amm.mmOraclePrice
+			.sub(oracleData.price)
+			.abs()
+			.mul(PERCENTAGE_PRECISION)
+			.div(BN.max(oracleData.price, ONE));
+		if (
+			!isOracleValid(
+				perpMarket,
+				oracleData,
+				stateAccountAndSlot.data.oracleGuardRails,
+				stateAccountAndSlot.slot
+			) ||
+			perpMarket.amm.mmOraclePrice.eq(ZERO) ||
+			perpMarket.amm.mmOracleSlot < oracleData.slot ||
+			pctDiff.gt(PERCENTAGE_PRECISION.divn(100)) // 1% threshold
+		) {
+			return { ...oracleData, fetchedWithMMOracle: true };
+		} else {
+			const conf = getOracleConfidenceFromMMOracleData({
+				mmOraclePrice: perpMarket.amm.mmOraclePrice,
+				mmOracleSlot: perpMarket.amm.mmOracleSlot,
+				oraclePriceData: oracleData,
+			});
+			return {
+				price: perpMarket.amm.mmOraclePrice,
+				slot: perpMarket.amm.mmOracleSlot,
+				confidence: conf,
+				hasSufficientNumberOfDataPoints: true,
+				fetchedWithMMOracle: true,
+				isMMOracleActive: oracleData.isMMOracleActive,
+			};
+		}
 	}
 
 	public getOracleDataForSpotMarket(marketIndex: number): OraclePriceData {
@@ -9215,7 +9266,8 @@ export class DriftClient {
 	) {
 		let feeTier;
 		const userHLM =
-			(user?.isHighLeverageMode() ?? false) || enteringHighLeverageMode;
+			(user?.isHighLeverageMode('Initial') ?? false) ||
+			enteringHighLeverageMode;
 		if (user && !userHLM) {
 			feeTier = user.getUserFeeTier(marketType);
 		} else {
@@ -9802,17 +9854,20 @@ export class DriftClient {
 			  })
 			: undefined;
 
-		const ix = await this.program.instruction.disableUserHighLeverageMode({
-			accounts: {
-				state: await this.getStatePublicKey(),
-				user,
-				authority: this.wallet.publicKey,
-				highLeverageModeConfig: getHighLeverageModeConfigPublicKey(
-					this.program.programId
-				),
-			},
-			remainingAccounts,
-		});
+		const ix = await this.program.instruction.disableUserHighLeverageMode(
+			false,
+			{
+				accounts: {
+					state: await this.getStatePublicKey(),
+					user,
+					authority: this.wallet.publicKey,
+					highLeverageModeConfig: getHighLeverageModeConfigPublicKey(
+						this.program.programId
+					),
+				},
+				remainingAccounts,
+			}
+		);
 
 		return ix;
 	}
@@ -9906,6 +9961,112 @@ export class DriftClient {
 			this.opts
 		);
 		return txSig;
+	}
+
+	public async updateMmOracleNative(
+		marketIndex: number,
+		oraclePrice: BN,
+		oracleSequenceId: BN
+	): Promise<TransactionSignature> {
+		const updateMmOracleIx = await this.getUpdateMmOracleNativeIx(
+			marketIndex,
+			oraclePrice,
+			oracleSequenceId
+		);
+
+		const tx = await this.buildTransaction(updateMmOracleIx, {
+			computeUnits: 5000,
+			computeUnitsPrice: 0,
+		});
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+
+		return txSig;
+	}
+
+	public async getUpdateMmOracleNativeIx(
+		marketIndex: number,
+		oraclePrice: BN,
+		oracleSequenceId: BN
+	): Promise<TransactionInstruction> {
+		const discriminatorBuffer = createNativeInstructionDiscriminatorBuffer(0);
+		const data = Buffer.alloc(discriminatorBuffer.length + 16);
+		data.set(discriminatorBuffer, 0);
+		data.set(oraclePrice.toArrayLike(Buffer, 'le', 8), 5); // next 8 bytes
+		data.set(oracleSequenceId.toArrayLike(Buffer, 'le', 8), 13); // next 8 bytes
+
+		// Build the instruction manually
+		return new TransactionInstruction({
+			programId: this.program.programId,
+			keys: [
+				{
+					pubkey: this.getPerpMarketAccount(marketIndex).pubkey,
+					isWritable: true,
+					isSigner: false,
+				},
+				{
+					pubkey: this.wallet.publicKey,
+					isWritable: false,
+					isSigner: true,
+				},
+				{
+					pubkey: SYSVAR_CLOCK_PUBKEY,
+					isWritable: false,
+					isSigner: false,
+				},
+				{
+					pubkey: await this.getStatePublicKey(),
+					isWritable: false,
+					isSigner: false,
+				},
+			],
+			data,
+		});
+	}
+
+	public async updateAmmSpreadAdjustmentNative(
+		marketIndex: number,
+		ammSpreadAdjustment: number
+	): Promise<TransactionSignature> {
+		const updateMmOracleIx = await this.getUpdateAmmSpreadAdjustmentNativeIx(
+			marketIndex,
+			ammSpreadAdjustment
+		);
+
+		const tx = await this.buildTransaction(updateMmOracleIx, {
+			computeUnits: 1000,
+			computeUnitsPrice: 0,
+		});
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+
+		return txSig;
+	}
+
+	public getUpdateAmmSpreadAdjustmentNativeIx(
+		marketIndex: number,
+		ammSpreadAdjustment: number // i8
+	): TransactionInstruction {
+		const discriminatorBuffer = createNativeInstructionDiscriminatorBuffer(1);
+		const data = Buffer.alloc(discriminatorBuffer.length + 4);
+		data.set(discriminatorBuffer, 0);
+		data.writeInt8(ammSpreadAdjustment, 5); // next byte
+
+		// Build the instruction manually
+		return new TransactionInstruction({
+			programId: this.program.programId,
+			keys: [
+				{
+					pubkey: this.getPerpMarketAccount(marketIndex).pubkey,
+					isWritable: true,
+					isSigner: false,
+				},
+				{
+					pubkey: this.wallet.publicKey,
+					isWritable: false,
+					isSigner: true,
+				},
+			],
+			data,
+		});
 	}
 
 	private handleSignedTransaction(signedTxs: SignedTxData[]) {