+++ sdk/src/math/oracles.ts
@@ -1,4 +1,11 @@
-import { AMM, OracleGuardRails, isVariant } from '../types';
+import {
+	AMM,
+	HistoricalOracleData,
+	OracleGuardRails,
+	OracleSource,
+	PerpMarketAccount,
+	isVariant,
+} from '../types';
 import { OraclePriceData } from '../oracles/types';
 import {
 	BID_ASK_SPREAD_PRECISION,
@@ -9,13 +16,8 @@ import {
 	FIVE_MINUTE,
 	PERCENTAGE_PRECISION,
 } from '../constants/numericConstants';
-import {
-	BN,
-	HistoricalOracleData,
-	OracleSource,
-	PerpMarketAccount,
-} from '../index';
 import { assert } from '../assert/assert';
+import { BN } from '@coral-xyz/anchor';
 
 export function oraclePriceBands(
 	market: PerpMarketAccount,